+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## Version 1.3.0
+
+- Re-structure extension
+- Add support per-workspace config
+
+## Version 1.2.0
+
+- Prevent remote document
+
+## Version 1.1.0
+
+- Optimized codes
+
 ## Version 1.0
 
-Initial release
+- Initial release